[🛠️ REFACTOR] handleGesture 훅으로 분리
